pyproject.toml
[project]
name = "leetcode-tracker"
version = "0.1.0"
license = { text = "MIT" }
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "certifi==2025.11.12",
    "click==8.3.1",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi==0.125.0",
    "fastapi-cli==0.0.16",
    "fastapi-cloud-cli==0.7.0",
    "fastar==0.8.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "jinja2==3.1.6",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "pyjwt>=2.10.1",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.21",
    "pyyaml==6.0.3",
    "rich==14.2.0",
    "rich-toolkit==0.17.1",
    "rignore==0.7.6",
    "sentry-sdk==2.48.0",
    "shellingham==1.5.4",
    "sqlmodel>=0.0.27",
    "starlette==0.50.0",
    "typer==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.2",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websockets==15.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]

[tool.ruff]
# The version of the Python interpreter you are targeting
target-version = "py313"
# The maximum length of a line before the formatter wraps it
line-length = 88

[tool.ruff.lint]
# "select" defines the rulesets to enable.
# E/W: Error and Warning (standard PEP 8)
# F: Pyflakes (logic errors)
# UP: pyupgrade (modernizes your syntax automatically)
# B: flake8-bugbear (catches likely bugs)
# I: isort (organizes your imports)
select = ["E", "F", "UP", "B", "I"]

# Rules to explicitly skip
ignore = [
    "E501", # Ignore line-length violations (the formatter handles this)
    "B008", # Allow function calls in default arguments (common in FastAPI)
]

[tool.ruff.lint.isort]
# Ensures imports are grouped logically
combine-as-imports = true
force-wrap-aliases = true

[tool.ruff.format]
# Use double quotes for strings, which is the standard for modern Python
quote-style = "double"
# Ensure indentation is consistent
indent-style = "space"

[tool.pyright]
venvPath = "."
venv = ".venv"
